// untitled folder/typings/types.ts
export type Board = {
  id: string;
  name: string;
  items_page: ItemsResponse;
  settings: Settings;
  weeklySchedules: WeeklySchedules;
};

export type Group = {
  id: string;
  title: string;
  items: Item[];
};

export type Column = {
  title: ColumnTitles;
  type: ColumnTypes;
  id: string;
  options?: columnOptions;
};

export type ItemsResponse = {
  cursor: string;
  items: Item[];
};

export type Item = {
  id: string;
  values: ColumnValue[];
  createdAt: number;
  completedAt?: number;
  status: ItemStatus;
  vertical?: boolean;
  visible: boolean;
  deleted: boolean;
  isScheduled?: boolean;
  shippingDetails: Address;
};
// ...
export type Settings = {
  automatronSettings: AutomatronSettings;
};

export type AutomatronSettings = PartialRecord<
  ColumnTitles,
  PartialRecord<ProgressStatus | ItemDesigns | ItemSizes, ItemStatus>
>;